Optimized drawing code by caching integer rotations of sprites
