+++ astroblast.py
@@ -10,6 +10,7 @@
 #   https://opengameart.org/content/rocks-ships-stars-gold-and-more
 
 from queue import Queue
+import cProfile
 
 import pygame
 
@@ -17,41 +18,52 @@ from dispatcher import Dispatcher
 from client import GameClient
 from server import GameServer
 
+
 def main():
     """Application entry point"""
-    dispatch_queue = Queue()
-    remote_queue = Queue()
-    # queue of messages for the server
-    server_queue = Queue()
-    # queue of messages for the client
-    local_client_queue = Queue()
+    profile = cProfile.Profile()
+    try:
+        # profile.enable()
 
-    dispatch = Dispatcher(dispatch_queue, server_queue, local_client_queue, remote_queue)
+        dispatch_queue = Queue()
+        remote_queue = Queue()
+        # queue of messages for the server
+        server_queue = Queue()
+        # queue of messages for the client
+        local_client_queue = Queue()
 
-    # Instantiate the server and client modules
-    server = GameServer(server_queue, dispatch)
-    client = GameClient(local_client_queue, dispatch_queue)
+        dispatch = Dispatcher(dispatch_queue, server_queue, local_client_queue, remote_queue)
 
-    dispatch.start()
-    # Start the server on a separate thread
-    server.start()
+        # Instantiate the server and client modules
+        server = GameServer(server_queue, dispatch)
+        client = GameClient(local_client_queue, dispatch_queue)
 
-    running = True
-    while running:
-        # Check & process incoming messages to the client
-        client.check_msgs()
-        # Update client-side game state / Process user input
-        client.update()
+        dispatch.start()
+        # Start the server on a separate thread
+        server.start()
 
-        # Shut down if the client quits
-        if not client.running:
-            running = False
-            # Make sure the server quits too
-            if server.running:
-                server.running = False
+        running = True
+        while running:
+            # Check & process incoming messages to the client
+            client.check_msgs()
+            # Update client-side game state / Process user input
+            client.update()
 
-    # Clean-up pyGame
-    print("Exiting...")
+            # Shut down if the client quits
+            if not client.running:
+                running = False
+                # Make sure the server quits too
+                if server.running:
+                    server.running = False
+
+        # Clean-up pyGame
+        print("Exiting...")
+
+        # profile.disable()
+    
+    finally:
+        # profile.print_stats('cumulative')
+        pass
 
 if __name__ == '__main__':
     main()
