+++ gamemodule.py
@@ -101,7 +101,7 @@ class GameModule():
 
     def log(self, msg):
         """Print a log message to stdout, along with the module's ID"""
-        print("%s(%d): %s" % (self.name, self.module_id, msg))
+        # print("%s(%d): %s" % (self.name, self.module_id, msg))
         pass
     
     def assert_msg_content(self, msg_type, msg_content, *expected_content):
